+++ config/settings/base.py
@@ -41,6 +41,7 @@ CONNECTIONS = {
         "PASSWORD": os.getenv("PASSWORD"),
         # APPS
         "PIPELINE_APPS": [
+            "wtfix.apps.utils.PipelineTerminationApp",
             "wtfix.apps.api.rest.RESTfulServiceApp",
             "wtfix.apps.brokers.RedisPubSubApp",
             "wtfix.apps.admin.HeartbeatApp",